+++ apps/astarte_realm_management_api/lib/astarte_realm_management_api_web/controllers/trigger_controller.ex
@@ -44,6 +44,21 @@ defmodule Astarte.RealmManagement.APIWeb.TriggerController do
         |> put_status(:conflict)
         |> render("already_installed_trigger.json")
 
+      {:error, :invalid_datastream_trigger} ->
+        conn
+        |> put_status(:bad_request)
+        |> render("invalid_datastream_trigger.json")
+
+      {:error, :unsupported_trigger_type} ->
+        conn
+        |> put_status(:bad_request)
+        |> render("unsupported_trigger_type.json")
+
+      {:error, :invalid_object_aggregation_trigger} ->
+        conn
+        |> put_status(:bad_request)
+        |> render("invalid_object_aggregation_trigger.json")
+
       # To FallbackController
       {:error, other} ->
         {:error, other}
@@ -53,6 +68,11 @@ defmodule Astarte.RealmManagement.APIWeb.TriggerController do
   def show(conn, %{"realm_name" => realm_name, "id" => id}) do
     with {:ok, trigger} <- Triggers.get_trigger(realm_name, id) do
       render(conn, "show.json", trigger: trigger)
+    else
+      {:error, :cannot_retrieve_simple_trigger} ->
+        conn
+        |> put_status(:internal_server_error)
+        |> render("cannot_retrieve_simple_trigger.json")
     end
   end
 
@@ -72,6 +92,11 @@ defmodule Astarte.RealmManagement.APIWeb.TriggerController do
     with {:ok, %Trigger{} = trigger} <- Triggers.get_trigger(realm_name, id),
          {:ok, %Trigger{}} <- Triggers.delete_trigger(realm_name, trigger) do
       send_resp(conn, :no_content, "")
+    else
+      {:error, :cannot_delete_simple_trigger} ->
+        conn
+        |> put_status(:internal_server_error)
+        |> render("cannot_delete_simple_trigger.json")
     end
   end
 end